Maximum number of orders constraint. (#18)
This PR makes sure the solutions produced by the token pair solver do not violate the maximum number of orders constraint.

Additionally:
* Adds unit testing for checking satisfiability of this constraint.
* Adds an example where the optimal exchange rate finder fails to detect the optimal solution due to one of this side constraints, namely the minimum tradable amount constraint.
* Adapts readme.

NOTE:
The code for finding the optimal xrate does not yet take this constraint into account (neither other side constraints such as the minimum tradable amount constraint). This means that with such side constraints, the solution might no longer be optimal. Trying to fix this requires some research, to be done in a future PR.