Config class (#36)
Refactors constants.py to make it possible to change them for unit testing.

Previously all constants were variables local to `src.core.constants`. When
one of these constants is imported to another module namespace, e.g.
```
from src.core.constants import MIN_TRADABLE_AMOUNT
```
a copy of the constant is created in the importing module. This means that
if the value of this constant is changed at runtime, which is useful for
non-intrusive unit tests, the new value won't be used.

The solution in the PR is to simply wrap all constants in a class. Since python
does not allow to import a specific class member, then the user needs to import
the full class. Therefore, any changes to one of the class (static) members will
be reflected throughout the code.

The wrapping class was named Config, and the constants.py file renamed accordingly.

This PR also makes sure there are no parameter defaults set to Config values, which is yet another way of shooting ourselves in the foot.