Economic viability constraint (#46)
Adds support for the economic viability constraint. The constraint is currently disabled (similarly to the standard solver), but can be enabled by setting "MIN_AVERAGE_ORDER_FEE" to some value >0.

About the implementation:

After chased some avenues along what I suggested here #14, I could not find a way to do this constraint in a way that it does not:

1. increase run time complexity and, at the same time,
2. guarantees that it finds an economic solution if there is an economic solution.

The current solution is very simple. It removes the most inviable order in the solution and tries to match everything again in a loop until it is economically viable.

It is a compromise. It violates (1) only when the constraint is active, i.e. the solution without the constraint is economic inviable. On the other hand the greedy removal approach still does not guarantee (2), but should be a good approximation in most cases.

Perhaps in the future we can revisit this an come up with a better idea.

Unit tests specifically for this constraint are going to be added in a subsequent PR.

Closes #14.