Fixed removing empty orders too early. (#49)
The previous PR added a statement to remove the orders with sell amount = 0 (empty orders). But this was being done before the final utility values were computed, which mean that they reflected the "touched orders" version of the objective, and not the "all orders" version we're optimizing for. This PR fixes it.