Mirror smart contract objective metrics (#54)
There is now 3 different ways to compute utiltiy/max utility/disregarded utility/objective.

1. The exact version using rational arithmetic (from Tom's paper)
2. The smart contract integer version
3. The version that tries to emulate the smart contract version in the standard solver, which was what was being used here as well to compute the true (integer) objective value of each solution and decide which token pair to match based on that.

This PR implements 2, keeping 3 around just in case. Once this is live, we can check in etherscan if the value computed by the smartcontract is the same as the one that is in the submitted json, thus validating if these 2 things finally converged.