Manual examples for unit tests (#56)
Unit testing in this repo is made using randomized testing. This procedure has been great so far, being able to catch a number of cases at least I would never consider doing manually. However it has the drawback of disabling something useful about unit tests, which is its ability to explain what a function does "by example", by reading the unit tests .

This PR tries to enable this here, by adding some explicit test cases in addition to the randomized unit tests. The test cases happened to be found by randomized unit testing, but that is irrelevant.

Note that this orthogonal to the existing e2e tests, since here we still test one "unit" at a time.

For some reason I can't recall, there is a small refactoring of the order.index, previously an integer, to an order.id which is now an hash.