Fixes smartcontract version of u, du. (#63)
A recent PR changed the way the integer version of the utility and disregarded utility is computed to mirror the smart contract version. Unfortunately, there was a bug there - the max_sell_amounts used in those computations should be the original and not those obtained after taking balances into account.

This PR fixes it. Using tenderly it is possible to confirm that finally the computation is in sync with the smart contract. Many thanks to @fleupold for teaching me how to use it.