Improves handling of economic viability constraint. (#66)
There are two improvements:

1. There was an obvious inefficiency as the algorithm was also removing orders with
zero buy amount, one at a time interleaved with solving.

2. Implements a test that checks if it is possible at all to arrive at an economic
viable solution by removing one order at a time, ignoring all other constraints (i.e.
assuming that utilities and prices would not change).

The runtimes here on by computer without these improvements on the instance above are:

- 1m:20s without iprovements
- 25s with improvement 1.
- 18s with improvement 2.

Currently both improvements are implemented, but I'm not sure if it is a good idea to keep
second, as it may exclude solutions and its impact is marginal. Keeping as it is.